better 64 bit support

+++ Utilities/interface/Types.h
@@ -14,7 +14,7 @@ typedef   short int              int16 ;
 typedef   unsigned char          size8 ; // 8 bit 0->255
 typedef   unsigned short int     size16; //16 bit 0->65,536
 typedef   unsigned int           size  ; //32 bit 0->4,294,967,296
-typedef   unsigned long long int size64; //64 bit 0->18,446,744,073,709,551,615
+typedef   uint64_t               size64; //64 bit 0->18,446,744,073,709,551,615
 
 
 typedef ROOT::Math::PtEtaPhiM4D<double>				                CylLorentzCoord;
